+++ resources/views/bootstrap-4/components/buttons/simple-button.blade.php
@@ -34,5 +34,5 @@
 </button>
 
 @if ($confirm !== null)
-    <x-modal-confirm :id="'confirm-modal-'.$confirmId" :title="trans('blade-ui-kit-bootstrap::modal.confirm')" />
+    <x-modal-confirm :id="'confirm-modal-'.$confirmId" :title="$confirmTitle" />
 @endif