+++ resources/views/bootstrap-5/components/buttons/simple-button.blade.php
@@ -42,5 +42,5 @@
 </button>
 
 @if ($confirm !== null)
-    <x-modal-confirm :id="'confirm-modal-'.$confirmId" :title="trans('blade-ui-kit-bootstrap::modal.confirm')" />
+    <x-confirm-modal :id="'confirm-modal-'.$confirmId" :title="$confirmTitle" />
 @endif