+++ src/Components/Buttons/HelpInfo.php
@@ -30,6 +30,7 @@ class HelpInfo extends BladeComponent
         public ?string $icon = null,
         public ?string $startIcon = null,
     ) {
+        $this->initAttributes();
         $this->onConstructing();
 
         $this->text ??= Str::ucfirst(trans('status.info'));