+++ src/Components/Modals/Confirm.php
@@ -4,18 +4,22 @@ declare(strict_types=1);
 
 namespace BladeUIKitBootstrap\Components\Modals;
 
-class Confirm extends Modal
+use BladeUIKitBootstrap\Components\BladeComponent;
+
+class Confirm extends BladeComponent
 {
+    public string $titleLabel;
+
     public function __construct(
-        string $id,
-        string $title,
-        bool $dismissable = true
+        public string $id,
+        public ?string $title = null,
     ) {
-        parent::__construct(
-            id: $id,
-            title: $title,
-            dismissable: $dismissable,
-        );
+        $this->initAttributes();
+        $this->onConstructing();
+
+        $this->titleLabel = str($id)->kebab()->append('-label')->toString();
+
+        $this->title ??= trans('blade-ui-kit-bootstrap::modal.confirm');
     }
 
     public function viewName(): string